The power model adapter is not needed in the kernel and references to it have been removed and replaced with the multiplicative inverse (reciprocal) of the performance model coefficients. Performance model coefficients are represented as 64-bit fixed point numbers with 48 bits after the fixed point.
Consider a possible implementation of VIC_timer. A VIC_timer has to be based on a real timer. Given a request for a VIC_timer to go off after some number of virtual instructions, a real timer has to be set up to go off after some time. Assuming that the MO remains constant over this time, the amount of time is equal to the instruction count divided by the performance coefficient of the current mode of operation.

To elliminate the need divide by fixed-point numbers in the kernel, the multiplicative inverse of the performance coefficients are also passed to the kernel by the allocator task. Division by one of the performance coefficients is equivalent to multiplying by the corresponding multiplicative inverse.