Modified scripts from LAMbS_VICbudget_test/expt_scripts/ to cycle through the full range of CPU frequencies during tests.
Added scripts with "test.csh" suffix to LAMbS_VICbudget_test/expt_scripts/ that run the same tests as the regular scripts but for a single alpha value and for a shorter duration.

Created a directory similar to LAMbS_VICbudget_test/ and LAMbS_VICtimer_test/ but for testing the invariance of VIC. Moved relevant csh and Octave scripts to the new folder.